+++ OuraRing/oura_ring_api.py
@@ -1,0 +1,82 @@
+import argparse
+import logging
+from typing import Any, Dict, List, Type, Union
+
+import requests
+
+logging.basicConfig(level=logging.INFO)
+
+# source: Mastering Object-Oriented Python - Second Edition, Steven F. Lott
+JSON = Union[Dict[str, Any], List[Any], int, str, float, bool, Type[None]]
+
+
+class OuraRing:
+    """This class encompasses a set of functions to connect to the Oura Ring
+    api and pull one's respective data. Note that this class assumes you are
+    accessing the api for personal use, and thus authorization is only via
+    personal access tokens.
+    """
+
+    def __init__(self, personal_access_token: str) -> None:
+        self.PAT = personal_access_token
+
+    def authenticate_api(self) -> None:
+        """Function takes a personal access token and connect to the oura ring
+        api. PATs can be generated by navigating to the following website and
+        creating an account: https://cloud.ouraring.com/personal-access-tokens
+
+        Docs: https://cloud.ouraring.com/docs/authentication
+        """
+        url = f"https://api.ouraring.com/v1/userinfo?access_token={self.PAT}"
+        response = requests.get(url)
+
+        if response.status_code == 200:
+            user_email = response.json()["email"]
+            logging.info("Response Code: %s", response.status_code)
+            logging.info("Oura account associated with email: %s", user_email)
+        else:
+            raise Exception(
+                f"Response Code: {response.status_code} - {response.reason}"
+            )
+
+    def get_daily_sleep_data(self, start_date: str, end_date: str) -> JSON:
+        """Function pulls Oura Ring daily sleep data for specified date range.
+
+        Docs:
+            https://cloud.ouraring.com/v2/docs#tag/Daily-Sleep
+
+        Arguments:
+            start_date: YYYY-MM-DD formatted start date
+            end_date: YYYY-MM-DD formatted end date
+
+        Returns:
+            JSON: json data of returned Oura Daily Sleep data for the specified
+                Oura user within a given timeframe. Example data can be found
+                here: https://cloud.ouraring.com/docs/sleep.
+        """
+        url = (
+            "https://api.ouraring.com/v2/usercollection/daily_sleep?"
+            f"access_token={self.PAT}"
+        )
+        params = {"start_date": start_date, "end_date": end_date}
+        headers = {"Authorization": f"Bearer {self.PAT}"}
+
+        response = requests.request("GET", url, headers=headers, params=params)
+        return response.json()
+
+
+if __name__ == "__main__":
+    # cli arguments
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-pat", "--pat", default=None, type=str)
+    options = parser.parse_args()
+
+    # authenticating api
+    oura = OuraRing(personal_access_token=options.pat)
+    oura.authenticate_api()
+
+    # get sleep data
+    start_date = "2022-09-01"
+    end_date = "2022-09-01"
+    sleep_data = oura.get_daily_sleep_data(start_date, end_date)
+    print(sleep_data)
